evaluate decimal expressions

+++ LL1Lib/Expression.h
@@ -19,6 +19,6 @@ public:
 };
 std::vector<std::string> convertTextToPostFix(const std::vector<std::string>& infixExpression);
 Expression TranslateWordsToTokens(std::vector<std::string> words);
-int evaluatePostFixExpression(const std::vector<std::string>& postFixExpression);
+double evaluatePostFixExpression(const std::vector<std::string>& postFixExpression);
 
 #endif //CS6820HW2_EXPRESSION_H
